Handle unknown metrics more gracefully
In the previous commit we fixed an issue where the Awair device started a
reporting a metric that wasn't listed in ExpecedMetrics. The symptom of this was
that the air-data endpoint starts 500ing. That's because Collect was returning a
metric with a descriptor that wasn't returned by Describe: the prometheus API
bails completely when this happens.

(I've actually just realised this is probably only because we're using
NewPedanticRegistry. So this commit could probably just have switched to
using a non-pedantic registry and left it at that. Oh well, it's the weekend,
I'm not gonna go back over this again)

With this commit we just log the missing metric and continue.